+++ Src/Roflcopter.Plugin.Tests/CopyFqnProviders/CustomSourceLinkFqnProviderTest.cs
@@ -20,10 +20,17 @@ namespace Roflcopter.Plugin.Tests.CopyFqnProviders
         [SetUp]
         public new void SetUp()
         {
-            _solutionPathProvider = new FixedSolutionPathProvider(TestDataPath2.Parent.Parent.Parent);
+#if RS20183
+            var testDataPath = TestDataPath2;
+#else
+            var testDataPath = TestDataPath;
+
+#endif
+
+            _solutionPathProvider = new FixedSolutionPathProvider(testDataPath.Parent.Parent.Parent);
 
             _gitRepositoryProvider = new TestGitRepositoryProvider();
-            _gitDirectoryPath = TestDataPath2.Parent.Combine(_gitRepositoryProvider.GitDirectoryName);
+            _gitDirectoryPath = testDataPath.Parent.Combine(_gitRepositoryProvider.GitDirectoryName);
 
             _gitDirectoryPath.CreateDirectory();
             _gitDirectoryPath.Combine("config").WriteAllText("https://repository.url/");