+++ Src/Roflcopter.Plugin/CopyFqnProviders/OptionsPages/CopyFqnProvidersOptionsPage.cs
@@ -4,8 +4,13 @@ using JetBrains.Annotations;
 using JetBrains.Application.UI.Options;
 using JetBrains.Application.UI.Options.Options.ThemedIcons;
 using JetBrains.Application.UI.Options.OptionsDialog;
-using JetBrains.DataFlow;
 using JetBrains.ReSharper.Features.Navigation.Options;
+#if RS20183
+using JetBrains.DataFlow;
+#else
+using JetBrains.Lifetimes;
+
+#endif
 
 namespace Roflcopter.Plugin.CopyFqnProviders.OptionsPages
 {
@@ -22,7 +27,11 @@ namespace Roflcopter.Plugin.CopyFqnProviders.OptionsPages
     {
         private const string PageId = nameof(CopyFqnProvidersOptionsPage);
 
-        public CopyFqnProvidersOptionsPage([NotNull] Lifetime lifetime, [NotNull] OptionsSettingsSmartContext optionsSettingsSmartContext) :
+        public CopyFqnProvidersOptionsPage(
+#if RS20183
+            [NotNull] 
+#endif
+            Lifetime lifetime, [NotNull] OptionsSettingsSmartContext optionsSettingsSmartContext) :
             base(lifetime, optionsSettingsSmartContext)
         {
             AddText("\"Copy Fully-qualified name/ Source browser URI to Clipboard\" will be extended by the following entries.");