+++ Src/Roflcopter.Plugin/TodoItems/OptionsPages/TodoItemsCountOptionsPage.cs
@@ -7,6 +7,10 @@ using JetBrains.Application.UI.Options.OptionsDialog.SimpleOptions;
 using JetBrains.DataFlow;
 using JetBrains.ReSharper.Features.Inspections.TodoItems;
 using JetBrains.ReSharper.Psi.Resources;
+#if !RS20183
+using JetBrains.Lifetimes;
+
+#endif
 
 namespace Roflcopter.Plugin.TodoItems.OptionsPages
 {
@@ -23,7 +27,11 @@ namespace Roflcopter.Plugin.TodoItems.OptionsPages
     {
         public const string OptionsPageId = nameof(TodoItemsCountOptionsPage);
 
-        public TodoItemsCountOptionsPage([NotNull] Lifetime lifetime, [NotNull] OptionsSettingsSmartContext optionsSettingsSmartContext) :
+        public TodoItemsCountOptionsPage(
+#if RS20183
+            [NotNull] 
+#endif
+            Lifetime lifetime, [NotNull] OptionsSettingsSmartContext optionsSettingsSmartContext) :
             base(lifetime, optionsSettingsSmartContext)
         {
             var enabledOption = AddBoolOption((TodoItemsCountSettings s) => s.IsEnabled, "Enabled");