+++ Src/Roflcopter.Plugin/UnitTesting/ParameterizedTestMissingParameterRemoveArgumentQuickFix.cs
@@ -7,6 +7,10 @@ using JetBrains.ReSharper.Psi.Tree;
 using JetBrains.TextControl;
 using JetBrains.Util;
 using ReSharperExtensionsShared.QuickFixes;
+#if !RS20183
+using JetBrains.Diagnostics;
+
+#endif
 
 namespace Roflcopter.Plugin.UnitTesting
 {