+++ Src/Roflcopter.Plugin/Utilities/AttributeExtensions.cs
@@ -17,7 +17,11 @@ namespace Roflcopter.Plugin.Utilities
                 {
                     return typeNamesToTest.Any(typeNameToTest =>
                     {
+#if RS20183
                         var interfaceTypeToTest = TypeFactory.CreateTypeByCLRName(typeNameToTest, attribute.PsiModule);
+#else
+                        var interfaceTypeToTest = TypeFactory.CreateTypeByCLRName(typeNameToTest, NullableAnnotation.Unknown, attribute.PsiModule);
+#endif
 
                         return typeElement.IsDescendantOf(interfaceTypeToTest.GetTypeElement());
                     });