+++ Src/Roflcopter.Plugin/VisualStudio/ModulesWindow/DebuggingControlIniActionBase.cs
@@ -5,6 +5,10 @@ using System.Diagnostics.CodeAnalysis;
 using System.Runtime.InteropServices;
 using JetBrains.Annotations;
 using JetBrains.Util;
+#if !RS20183
+using JetBrains.Diagnostics;
+
+#endif
 
 namespace Roflcopter.Plugin.VisualStudio.ModulesWindow
 {