+++ History.md
@@ -1,4 +1,5 @@
-### vNext ###
+### 1.6.0 ###
+- ReSharper and Rider 2019.1 support
 - Updated default "allowed postfix regex" for 'MismatchedFileName' warning
 
 ### 1.5.0 ###